Fix(catalog): Only remove metadata files if TableOp created a new one
In multiple Table ops (Glue, Dynamo, BigQuery and Nessie) a failing
operation will clean up the metadata file.
However it should only do so if it was the one to create the metadata file
in the first place.

Otherwise, an operation like `registerTable` could corrupt other catalog that
rely on the existence of the metadata file

+++ aws/src/main/java/org/apache/iceberg/aws/dynamodb/DynamoDbTableOperations.java
@@ -152,8 +152,10 @@ class DynamoDbTableOperations extends BaseMetastoreTableOperations {
       }
     } finally {
       try {
-        if (commitStatus == CommitStatus.FAILURE) {
-          // if anything went wrong, clean up the uncommitted metadata file
+        if (commitStatus == CommitStatus.FAILURE
+            && !Objects.equals(newMetadataLocation, metadata.metadataFileLocation())) {
+          // if anything went wrong, clean up the uncommitted metadata file, if a new one was
+          // created
           io().deleteFile(newMetadataLocation);
         }
       } catch (RuntimeException e) {
